Move stage2 classes into their own files
