+++ src/arm64/stage2_build_tape.h
@@ -1,5 +1,5 @@
-#ifndef SIMDJSON_ARM64_STAGE2_BUILD_TAPE_H
-#define SIMDJSON_ARM64_STAGE2_BUILD_TAPE_H
+#ifndef SIMDJSON_ARM64_STAGE2_H
+#define SIMDJSON_ARM64_STAGE2_H
 
 #include "simdjson.h"
 #include "arm64/implementation.h"
@@ -10,10 +10,11 @@ namespace simdjson {
 namespace arm64 {
 
 #include "generic/atomparsing.h"
-#include "generic/stage2_build_tape.h"
-#include "generic/stage2_streaming_build_tape.h"
+#include "generic/structural_iterator.h"
+#include "generic/structural_parser.h"
+#include "generic/streaming_structural_parser.h"
 
 } // namespace arm64
 } // namespace simdjson
 
-#endif // SIMDJSON_ARM64_STAGE2_BUILD_TAPE_H
+#endif // SIMDJSON_ARM64_STAGE2_H
