+++ src/fallback/stage1.h
@@ -1,5 +1,5 @@
-#ifndef SIMDJSON_FALLBACK_STAGE1_FIND_MARKS_H
-#define SIMDJSON_FALLBACK_STAGE1_FIND_MARKS_H
+#ifndef SIMDJSON_FALLBACK_STAGE1_H
+#define SIMDJSON_FALLBACK_STAGE1_H
 
 #include "simdjson.h"
 #include "fallback/implementation.h"
@@ -211,4 +211,4 @@ WARN_UNUSED error_code implementation::minify(const uint8_t *buf, size_t len, ui
 } // namespace fallback
 } // namespace simdjson
 
-#endif // SIMDJSON_FALLBACK_STAGE1_FIND_MARKS_H
+#endif // SIMDJSON_FALLBACK_STAGE1_H
