+++ src/fallback/stage2.h
@@ -1,5 +1,5 @@
-#ifndef SIMDJSON_FALLBACK_STAGE2_BUILD_TAPE_H
-#define SIMDJSON_FALLBACK_STAGE2_BUILD_TAPE_H
+#ifndef SIMDJSON_FALLBACK_STAGE2_H
+#define SIMDJSON_FALLBACK_STAGE2_H
 
 #include "simdjson.h"
 
@@ -11,10 +11,11 @@ namespace simdjson {
 namespace fallback {
 
 #include "generic/atomparsing.h"
-#include "generic/stage2_build_tape.h"
-#include "generic/stage2_streaming_build_tape.h"
+#include "generic/structural_iterator.h"
+#include "generic/structural_parser.h"
+#include "generic/streaming_structural_parser.h"
 
 } // namespace fallback
 } // namespace simdjson
 
-#endif // SIMDJSON_FALLBACK_STAGE2_BUILD_TAPE_H
+#endif // SIMDJSON_FALLBACK_STAGE2_H
