+++ src/generic/streaming_structural_parser.h
