+++ src/stage2_build_tape.cpp
@@ -12,14 +12,14 @@ void found_bad_string(const uint8_t *buf);
 #endif
 
 #if SIMDJSON_IMPLEMENTATION_ARM64
-#include "arm64/stage2_build_tape.h"
+#include "arm64/stage2.h"
 #endif
 #if SIMDJSON_IMPLEMENTATION_FALLBACK
-#include "fallback/stage2_build_tape.h"
+#include "fallback/stage2.h"
 #endif
 #if SIMDJSON_IMPLEMENTATION_HASWELL
-#include "haswell/stage2_build_tape.h"
+#include "haswell/stage2.h"
 #endif
 #if SIMDJSON_IMPLEMENTATION_WESTMERE
-#include "westmere/stage2_build_tape.h"
+#include "westmere/stage2.h"
 #endif
